+++ src/ShaderPlayground.Core/CompilerNames.cs
@@ -11,6 +11,7 @@
         public const string IntelShaderAnalyzer = "intelshaderanalyzer";
         public const string Metal = "metal";
         public const string MetalLib = "metallib";
+        public const string MetalShaderConverter = "metal-shader-converter";
         public const string Naga = "naga";
         public const string PowerVR = "powervr";
         public const string Rga = "rga";